% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futureverse.R
\docType{package}
\name{futureverse-package}
\alias{futureverse}
\alias{futureverse-package}
\title{futureverse: Install 'Futureverse' in One Go}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The 'Futureverse' is a set of packages for parallel and distributed processing with the 'future' package at its core, cf. Bengtsson (2021) \doi{10.32614/RJ-2021-048}. This package is designed to make it easy to install common 'Futureverse' packages in a single step. This package is intended for end-users, interactive use, and R scripts. Packages must not list it as a dependency - instead, explicitly declare each 'Futureverse' package as a dependency as needed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.futureverse.org}
}

}
\author{
\strong{Maintainer}: Henrik Bengtsson \email{henrikb@braju.com} (\href{https://orcid.org/0000-0002-7579-5165}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Henrik Bengtsson \email{henrikb@braju.com} (\href{https://orcid.org/0000-0002-7579-5165}{ORCID}) [copyright holder]
}

}
\keyword{internal}
