% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futureverse_packages.R
\name{futureverse_packages}
\alias{futureverse_packages}
\title{List all packages in the Futureverse}
\usage{
futureverse_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include \strong{futureverse} in the list?}
}
\description{
List all packages in the Futureverse
}
\examples{
futureverse_packages()

}
