% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futureverse_update.R
\name{futureverse_update}
\alias{futureverse_update}
\title{Update Futureverse packages}
\usage{
futureverse_update(recursive = FALSE, repos = getOption("repos"))
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
tidyverse packages.}

\item{repos}{The repositories to use to check for updates.
Defaults to \code{getOption("repos")}.}
}
\value{
Returns nothing; outputs installation instructions, if needed.
}
\description{
This will check to see if all Futureverse packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
futureverse_update()
}

}
