% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futureverse.R
\docType{package}
\name{futureverse-package}
\alias{futureverse}
\alias{futureverse-package}
\title{futureverse: Easily Install and Load the 'Futureverse'}
\description{
The 'futureverse' is a set of packages for parallel and distributed process with the 'future' package at its core. This package is designed to make it easy to install and load multiple 'futureverse' packages in a single step. Learn more about the 'futureverse' at \url{https://www.futureverse.org}. This package is intended for end-users and R scripts. Packages must not list it as a dependency - instead, declare each Futureverse package needed.
}
\author{
\strong{Maintainer}: Henrik Bengtsson \email{henrikb@braju.com} [copyright holder]

}
\keyword{internal}
